refactor : 각 계층을 패키지별로 관리하도록 수정
