+++ src/main/java/roomescape/time/controller/TimeController.java
@@ -1,4 +1,4 @@
-package roomescape.time;
+package roomescape.time.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
