+++ src/test/java/roomescape/MissionStepTest.java
@@ -24,8 +24,8 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.reservation.Reservation;
-import roomescape.reservation.ReservationController;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.controller.ReservationController;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
