+++ src/test/java/roomescape/time/controller/TimeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.time.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -9,17 +9,18 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
+import roomescape.time.controller.TimeController;
 
-@DisplayName("예약 view 컨트롤러")
-@WebMvcTest(ReservationController.class)
-class ReservationControllerTest {
+@DisplayName("시간 view 컨트롤러")
+@WebMvcTest(TimeController.class)
+class TimeControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
     @Test
-    void homeViewTest() throws Exception {
-        mockMvc.perform(get("/reservation"))
+    void timeViewTest() throws Exception {
+        mockMvc.perform(get("/time"))
                 .andExpect(status().isOk())
-                .andExpect(view().name("admin/reservation"));
+                .andExpect(view().name("admin/time"));
     }
 }
